Update ISSLocation.java
timestamp should be int or maybe long (int should work until 2035 or something like that), but for sure not double

+++ src/main/java/iss/model/ISSLocation.java
@@ -30,7 +30,7 @@ public class ISSLocation {
         this.longitude = longitude;
     }
 
-    public double getTimestamp() {
+    public int getTimestamp() {
         return timestamp;
     }
